requirements.txt
annotated-types==0.7.0
anyio==4.10.0
click==8.2.1
dnspython==2.7.0
email-validator==2.3.0
fastapi==0.116.1
h11==0.14.0
idna==3.10
pydantic==2.11.7
pydantic_core==2.33.2
sniffio==1.3.1
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
openai==1.51.0
python-multipart==0.0.12

# Database dependencies
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# JWT and Auth
PyJWT==2.8.0
python-jose[cryptography]==3.3.0

# HTTP client for API calls
httpx==0.25.0

# Environment management
python-dotenv==1.0.0

# Google Cloud Pub/Sub for event emission
google-cloud-pubsub==2.18.4
